Update testcases with right xlen.

+++ gcc/testsuite/gcc.target/riscv/rvp64/builtin-rvp64-insb.c
@@ -1,7 +1,7 @@
 /* insb also appears on filename, so scan-assembler-times plus 1 */
 /* This is a test program for insb instruction.  */
 /* { dg-do compile { target riscv64*-*-* } } */
-/* { dg-options "-march=rv64gc_zpn_zpsf -mabi=lp64d -O0" } */
+/* { dg-options "-march=rv64gc_zpn -mabi=lp64d -O0" } */
 
 
 #include <rvp_intrinsic.h>
@@ -9,7 +9,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-uint64_t bsni (uint64_t ra, uint32_t rb)
+uint64_t bsni (uint32_t ra, uint32_t rb)
 {
   return __rv_insb (ra, rb, 1);
 }
